Robust checks in build.py

+++ build.py
@@ -9,13 +9,43 @@ with file( 'templates/footer.tpl' ) as f:
 def getHeader( classname ):
     return header.replace( 'BODYCLASSNAME', classname )
 
+def notInstalled( program ):
+    return not isInstalled( program )
+
+def isInstalled( program ):
+    cmd = 'type %s > /dev/null 2>&1' % program
+    p = subprocess.Popen( cmd, shell=True )
+    return p.wait() == 0
+
+def warning( text ):
+    # http://stackoverflow.com/questions/287871/print-in-terminal-with-colors-using-python
+    WARNING = '\033[93m'
+    ENDC = '\033[0m'
+    print '%sWARNING: %s%s' % ( WARNING, text, ENDC )
+
 def buildJs():
+    if notInstalled( 'r.js' ):
+        warning( 'r.js not installed. Not building javascript files' )
+        return
+
     cmds = [ 'rm -rf build/*', 'r.js -o build.js' ]
     for cmd in cmds:
         p = subprocess.Popen( cmd.split( ' ' ))
         p.wait()
 
-def buildCss():
+def buildScss():
+    if notInstalled( 'scss' ):
+        warning( 'scss not installed. Not building scss styles' )
+        return
+    cmd = 'scss style/style.scss:style/style.css'
+    p = subprocess.Popen( cmd.split( ' ' ))
+    p.wait()
+
+
+def compactCss():
+    if notInstalled( 'r.js' ):
+        warning( 'r.js not installed. Not compressing/coying css files' )
+        return
     cmd = 'r.js -o cssIn=style/style.css out=build/style.css'
     p = subprocess.Popen( cmd.split( ' ' ))
     p.wait()
@@ -37,8 +67,7 @@ def buildPage( filename ):
 def pack():
     cmds = [
         'rm packed.tar.gz',
-        'cd build; tar cvzf packed.tar.gz *; mv packed.tar.gz ..',
-        'tar -tvf packed.tar.gz'
+        'cd build; tar cvzf packed.tar.gz * >/dev/null; mv packed.tar.gz ..',
     ]
     for cmd in cmds:
         p = subprocess.Popen( cmd, shell=True )
@@ -56,7 +85,8 @@ def copyImages():
 if __name__ == '__main__':
     if len( sys.argv ) <= 2 or sys.argv[ 1 ] != 'tpl':
         buildJs()
-        buildCss()
+        buildScss()
+        compactCss()
     buildPages()
     copyImages()
     pack()
